\name{GeneSOM-internal}
\alias{ciplot}
\alias{inrange}
\alias{plotcell}
\alias{somgrids}
\alias{somsum}
\alias{som.par}
\title{Internal GeneSOM functions}
\description{
  Internal GeneSOM functions.
}
\usage{
ciplot(x, y, se, n=1, d, ywindow)
inrange(x, xlim)
plotcell(x, y, dat, code, n, sdbar=1, ylim, yadj)
somgrids(xdim, ydim, color, yadj=0.1, hexa, ntik, ylim)
somsum(obj)
som.par(obj)
}
\details{
  These are not to be called directly by the user.
}
\keyword{internal}
