\name{summary.som}
\alias{summary.som}
\alias{summary.print.som}
\alias{print.som}
\title{ summarize a som object }
\description{print out the configuration parameters of a som object
}
\usage{
\method{summary}{som}(obj)
\method{print}{som}(obj)
}

\arguments{
  \item{obj}{ a `som' object }
}

\author{
  Jun Yan <jyan@stat.wisc.edu>
}

\keyword{print}
