\name{yeast}
\alias{yeast}
\title{ yeast cell cycle }
\usage{
data(yeast)
}
\description{
The \code{yeast} data frame has 6601 rows and 16 columns, i.e., 6601
genes, measured at 16 time points.
}
\format{
This data frame contains the following columns:
\describe{
\item{zero}{
a numeric vector
}
\item{ten}{
a numeric vector
}
\item{twenty}{
a numeric vector
}
\item{thirty}{
a numeric vector
}
\item{fourty}{
a numeric vector
}
\item{fifty}{
a numeric vector
}
\item{sixty}{
a numeric vector
}
\item{seventy}{
a numeric vector
}
\item{eighty}{
a numeric vector
}
\item{hundred}{
a numeric vector
}
\item{one.ten}{
a numeric vector
}
\item{one.twenty}{
a numeric vector
}
\item{one.thirty}{
a numeric vector
}
\item{one.fourty}{
a numeric vector
}
\item{one.fifty}{
a numeric vector
}
\item{one.sixty}{
a numeric vector
}
}
}
\source{
http://genomics.stanford.edu
}
\examples{
data(yeast)
}
\keyword{datasets}
