\name{GeneSOM-internal}
\alias{ciplot}
\alias{inrange}
\alias{plotcell}
\alias{somgrids}
\title{ Internal GeneSOM functions}
\description{
  Internal GeneSOM functions.
}
\usage{
ciplot(x, y, se, n=1, d, ywindow)
inrange(x, xlim)
plotcell(x, y, dat, code, n, sdbar=1, ylim, yadj)
somgrids(xdim, ydim, color, yadj=0.1, hexa, ntik, ylim)
}
\details{
  These are not to be called by the user.
}
\keyword{
  utilities
}
