\name{summary.som}
\alias{summary.som}
\alias{print.som}
\title{ summarize a som object }
\description{print out the configuration parameters of a som object
}
\usage{
\method{summary}{som}(object, ...)
\method{print}{som}(x, ...)
}

\arguments{
  \item{object, x}{ a `som' object }
}

\author{
  Jun Yan <jyan@stat.wisc.edu>
}

\keyword{print}

